+++ public/js/controllers/borrowController.js
@@ -65,7 +65,7 @@ angular.module('BibBox').controller('BorrowController', ['$scope', '$controller'
       $scope.baseResetIdleWatch();
 
       // Check if all tags in series have been added.
-      if (!material.invalid && !material.loading && (!material.success || material.status === 'error') && $scope.allTagsInSeries(material)) {
+      if (!material.invalid && !material.loading && !material.success && $scope.allTagsInSeries(material)) {
         // If a tag is missing from the device.
         if ($scope.anyTagRemoved(material.tags)) {
           material.tagRemoved = true;